OCR_Form_extraction/requirements.txt
azure-ai-documentintelligence>=1.0.0b4
azure-identity>=1.12.0
python-dotenv>=0.19.0
streamlit>=1.22.0
pytest>=7.0.0
pydantic>=1.10.0
openai>=1.12.0
matplotlib>=3.5.0
jinja2>=3.0.0